Attempt to use tsx again
